part of SCRUM-52

+++ Assets/Scripts/Player/PlayerDefense.cs
@@ -26,7 +26,7 @@ public class PlayerDefense : MonoBehaviour
         }
     }
 
-    public int ApplyDefense(int incomingDamage) 
+    public float ApplyDefense(float incomingDamage) 
     {
         if (!isDefending)
             return incomingDamage;
